mv func to sp_utils
